Initial and basic support for module streams.
We need some more contributions to make module config files added to
the etc_mock/rocky.*/modules for each of the modules and branches
found at: https://git.centos.org/projects/modules/%2A

In each of those directory branches you will find a YAML file that
shows the RPM Macros that need to be added to each config, plus the
module name(s) needed for builds. Use the perl and javascript-tools
examples that are there for guidance.